added sampleIds as option to the hierarchy service
