+++ persistence/persistence-api/src/main/java/org/cbioportal/persistence/GenesetHierarchyRepository.java
@@ -2,11 +2,14 @@ package org.cbioportal.persistence;
 
 import java.util.List;
 
-import org.cbioportal.model.GenesetAlteration;
+import org.cbioportal.model.Geneset;
 import org.cbioportal.model.GenesetHierarchyInfo;
 
 public interface GenesetHierarchyRepository {
 
-    List<GenesetHierarchyInfo> getGenesetHierarchyItems(String geneticProfileId, List<GenesetAlteration> genesetScores, 
-    		Integer percentile, Double scoreThreshold, Double pvalueThreshold);
+	List<GenesetHierarchyInfo> getGenesetHierarchyParents(List<String> genesetIds);
+
+	List<Geneset> getGenesetHierarchyGenesets(Integer nodeId);
+
+	List<GenesetHierarchyInfo> getGenesetHierarchySuperNodes(List<String> genesetIds);
 }
