+++ service/src/main/java/org/cbioportal/service/GenesetHierarchyService.java
@@ -7,6 +7,12 @@ import org.cbioportal.service.exception.GeneticProfileNotFoundException;
 
 public interface GenesetHierarchyService {
 
-	List<GenesetHierarchyInfo> getGenesetHierarchyInfo(String geneticProfileId, Integer percentile, Double scoreThreshold, Double pvalueThreshold) throws GeneticProfileNotFoundException;
+	List<GenesetHierarchyInfo> fetchGenesetHierarchyInfo(String geneticProfileId, Integer percentile, Double scoreThreshold, Double pvalueThreshold) throws GeneticProfileNotFoundException;
+
+	List<GenesetHierarchyInfo> fetchGenesetHierarchyInfo(String geneticProfileId, Integer percentile, Double scoreThreshold,
+			Double pvalueThreshold, List<String> sampleIds) throws GeneticProfileNotFoundException;
+
+	List<GenesetHierarchyInfo> fetchGenesetHierarchyInfo(String geneticProfileId, Integer percentile, Double scoreThreshold,
+			Double pvalueThreshold, String sampleListId) throws GeneticProfileNotFoundException;
 
 }
