+++ web/src/main/java/org/cbioportal/web/GenesetHierarchyController.java
@@ -14,6 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -32,7 +33,8 @@ public class GenesetHierarchyController {
     @Autowired
     private GenesetHierarchyService genesetHierarchyService;
 
-    @RequestMapping(value = "/genesets/hierarchy", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    @RequestMapping(value = "/genesets/hierarchy/fetch", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation("Get gene set hierarchical organization information. I.e. how different gene sets relate to other gene sets, in a hierarchy")
     public ResponseEntity<List<GenesetHierarchyInfo>> fetchGenesetHierarchyInfo(
 		@ApiParam(required = true, value = "Genetic Profile ID e.g. brca_tcga_gsva_scores. The final hierarchy "
@@ -49,10 +51,27 @@ public class GenesetHierarchyController {
     	@ApiParam("p-value threshold. Filters out gene sets for which the score p-value is higher than this threshold.")
         @Max(1)
         @Min(0)
-        @RequestParam(defaultValue = "0.05") Double pvalueThreshold)
+        @RequestParam(defaultValue = "0.05") Double pvalueThreshold,
+        @ApiParam(required = false, value = "Identifier of pre-defined sample list with samples to query, e.g. brca_tcga_all")
+    	@RequestParam(required = false) String sampleListId,
+        @ApiParam(required = false, value = "Fill this one if you want to specify a subset of samples:"
+        		+ " sampleIds: custom list of samples or patients to query, e.g. [\"TCGA-A1-A0SD-01\", \"TCGA-A1-A0SE-01\"]")
+        @RequestBody(required = false) List<String> sampleIds)
         throws GeneticProfileNotFoundException {
 
-        return new ResponseEntity<>(
-        		genesetHierarchyService.getGenesetHierarchyInfo(geneticProfileId, percentile, scoreThreshold, pvalueThreshold), HttpStatus.OK);
+	        if (sampleListId != null && sampleListId.trim().length() > 0) {
+	    		return new ResponseEntity<>(
+	    				genesetHierarchyService.fetchGenesetHierarchyInfo(geneticProfileId, percentile, scoreThreshold, pvalueThreshold, sampleListId)
+		        		, HttpStatus.OK);
+	    	} else if (sampleIds != null && sampleIds.size() > 0){
+		        return new ResponseEntity<>(
+		        		genesetHierarchyService.fetchGenesetHierarchyInfo(geneticProfileId, percentile, scoreThreshold, pvalueThreshold, sampleIds)
+		        		, HttpStatus.OK);
+	    	} else {
+		        return new ResponseEntity<>(
+		        		genesetHierarchyService.fetchGenesetHierarchyInfo(geneticProfileId, percentile, scoreThreshold, pvalueThreshold)
+		        		, HttpStatus.OK);
+	    	}
+        
     }
 }
